+++ CHANGELOG.md
@@ -11,6 +11,17 @@ All notable changes to this project will be documented in this file.
 + create scores
 + tower shoot units
 + units movements limited by borders
++ add color for sides
+
+## 0.1.1.1 - 2020-11-26
+
+### Changed
+
++ middle width is 3 instead of 1 to allocate two different unit spawns
+
+### Added
+
++ random unit spawn for both sides
 
 ## 0.1.1 - 2020-11-26